release: initial commit of working extension
This is the initial commit of all work made to get unearth from concept
to reality. Currently in beta and unreleased in any extension store
(Firefox AMO or Chrome Webstore), this version works for Firefox.

A few todos are needed before a public release, but this version is pr-
etty close to the launch product.