fix: add output file path to `npm sass` script
Problem:
After removing `main.css` from git and running the `npm start` or `npm
sass` commands, there would be no custom styling applied to the new tab
page as expected. This was due to the `npm sass` command (which converts
the `main.scss` to `main.css`) missing an output argument.

Solution:
Adding a second argument to the `npm sass` command for the output file-
path correctly generates the `main.css` file needed for the extension.