update site to utilize nuxt layouts and pages
Initially was just toying around / hacking to get a landing page setup
using Nuxt. After deciding to continue using it, I decided to clean up
the source code a bit and utilize the Nuxt layouts and pages directories.

This has made the code much more modular and easier to deal with.