Wire syncing saves with Redux; success step added
Hooked up the `getThings` action creator with the Sync step of the
onboarding process. Also threw in basic error handling in case the
sync process goes awry.

Built a simple success component as well as part of the last step in
the onboarding flow with a message telling users to watch their email
for their first newsletter.

Todos:

 - refactor sync step to show when successful but don't immediately
   show the success component; instead, have a submit button that
   appears that will POST our backend API with the data collected from
   the onboarding process

 - build out necessary backend endpoints for the above todo

 - error handling; styling (kinda like the minimalistic feel lol)