Fix: use thing permalink rather than URL
Over the past few days, I've noticed that some things in the newsletter
will not be hyperlinked. Further investigation shows that this happens
specifically for comments since they do not have the `url` property.

To standardize/normalize what's expected when clicking a hyperlink in
newsletter, I've decided to just link all things (posts and comments)
using their `permalink` property. This way, every link clicked takes
the user to Reddit which will encourage them to possibly unsave the
thing if they no longer find value in it.