Fix: use `link_title` for comments
Previously, the code assumed there was a `title` property for all types
of Reddit things (posts and comments). This is not true--`title` only
exists for things of type posts and therefore when a user's saves were
syned with Unearth, all saved comments had missing titles.

This was fixed by evaluating whether or not the `title` property existed
and if not, fallback to the `link_title`. This means posts will still use
the proper title and now comments will use the title of their parent post.