Fix: ignore fetching things with empty/null titles
This is a band-aid of sorts for those things already in production.
Although the previous commit (2f66c87) prevents from adding new things
without a title, this commit is necessary as there are already things
with empty/null titles.

This commit makes it so that those users already receiving emails
will no longer receive them with missing things as has already
happened.