Initial stab at refactoring mailing service
This commit includes the changes for the refactored mailing service.
This refactored form uses abstractions to reduce the amount of redundant
code previously introduced when said service was originally implemented.

This commit also includes the updated jobs that use the new mailers, as
well as the updated and new email templates (newsletter and welcome emails).