Fix: pass email directly to mailer; add error handling
The `sendWelcomeEmail` mailer was not sending emails to new users due
to missing email addresses. Previously, the PUT user controller would
query for the user with the given id. The code then assumed there was
an email address associated with said user and would pass `user.email`
to the `sendWelcomeEmail` function. Because the user query occurred
before updating the user's info (adding their email address and freq-
unecy preference), `user.email` was null and thus caused no emails to
be sent.

The solution is to instead pass the email coming from the `req.body`
directly to the `sendWelcomeEmail` function.