Implement ability to unsubscribe from newsletters
This commit makes it possible for users to unsubscribe/stop receiving
emails. This comes in handy when all of the user's saved things have
been surfaced and they no longer wish to receive the empty emails.

Note: the actual 'unsubscribe' option in the users table frequency column
was updated via a new migration that will be landing via #17 (specifically
https://github.com/cedricium/unearth/commit/192e66e5c4892213cd23f02789f4cf3a700c5c59#diff-2f9bef2e2d1c597f70787dd298e5e0aa)