Fix: use base64url encoding when generating unsubscribe hash
Problem:
A user reported haivng issues when trying to use the unsubscribe link
included in the updated emails. After investigating, I noticed the
mismatch between the hash in unsubscribe link and the query parameter
being sent/interpreted on the backend: one had a '+' and the other had
a ' ' (space) instead.

Solution:
As outlined in the test case description, a little extra work is needed
to ensure the encoded hashes are URL-safe. This is accomplished using
base64url encoding which handles the '+', '/', and '=' characters for us.

References:
 - https://en.wikipedia.org/wiki/Base64#URL_applications